fixed search issue

+++ application/controllers/Profile.php
@@ -33,9 +33,10 @@ class Profile extends CI_Controller {
 		//all followings
 		$followings = $this->db->query('SELECT COUNT(*) As total from follow_master where user_id = '.$user_id)->row_array();
 
-		$friends_to = $this->db->query('SELECT *, a.id As peer_master_id from peer_master As a INNER JOIN user As b ON a.peer_id = b.id WHERE a.user_id = '.$user_id.' AND (a.status = 2) ORDER BY a.id DESC')->result_array();
+		/*$friends_to = $this->db->query('SELECT *, a.id As peer_master_id from peer_master As a INNER JOIN user As b ON a.peer_id = b.id WHERE a.user_id = '.$user_id.' AND (a.status = 2) ORDER BY a.id DESC')->result_array();
 		$friends_from = $this->db->query('SELECT *, a.id As peer_master_id  from peer_master As a INNER JOIN user As b ON a.user_id = b.id WHERE a.peer_id = '.$user_id.' AND (a.status = 2) ORDER BY a.id DESC')->result_array();
-		$peer_to = array_merge($friends_to, $friends_from);
+		$peer_to = array_merge($friends_to, $friends_from);*/
+		$peer_to = $this->db->query('SELECT *, a.id As friends_id from friends As a INNER JOIN user As b ON a.peer_id = b.id WHERE a.user_id ='.$user_id)->result_array();
 		$peer_from = $this->db->query('SELECT *, c.id As notify_id, a.id As action_id from peer_master As a INNER JOIN user As b ON a.user_id = b.id INNER JOIN notification_master As c ON a.id = c.action_id WHERE a.peer_id = '.$user_id.' AND (a.status = 1) ORDER BY a.id DESC')->result_array();
 		$data['all_connections'] = $peer_to;
 		$data['all_requests'] = $peer_from;
@@ -339,18 +340,15 @@ class Profile extends CI_Controller {
 	public function unfriend()
 	{
 		if($this->input->post()){
-
-			$peer_master_id    = $this->input->post('peer_master_id');
-			$detail = $this->db->get_where($this->db->dbprefix('peer_master'), array('id'=> $peer_master_id))->row_array();
+			$friends_id    = $this->input->post('friends_id');
+			$friend_detail = $this->db->query('SELECT * from friends WHERE id = '.$friends_id)->row_array();
 			//delete from friends table
-			$this->db->where(array('user_id' => $detail['user_id'], 'peer_id' => $detail['peer_id']));
+			$this->db->where(array('user_id' => $friend_detail['user_id'], 'peer_id' => $friend_detail['peer_id']));
 			$this->db->delete('friends');
-			$this->db->where(array('user_id' => $detail['peer_id'], 'peer_id' => $detail['user_id']));
+			$this->db->where(array('user_id' => $friend_detail['peer_id'], 'peer_id' => $friend_detail['user_id']));
 			$this->db->delete('friends');
-			//delete from peer master table
-			$this->db->where('id', $peer_master_id);
-			$this->db->delete('peer_master');
-
+			$this->db->where(array('user_id'=> $friend_detail['user_id'], 'peer_id' => $friend_detail['peer_id']))->delete('peer_master');
+			$this->db->where(array('peer_id'=> $friend_detail['user_id'], 'user_id' => $friend_detail['peer_id']))->delete('peer_master');
 			redirect(site_url('Profile/timeline?tab=peers'));
 		}
 	}
